Move/Pin to 3.1 runtime (#688)
Move/Pin to 3.1 runtime

+++ test/coverlet.integration.tests/BaseTest.cs
@@ -25,6 +25,12 @@ namespace Coverlet.Integration.Tests
         private BuildConfiguration GetAssemblyBuildConfiguration()
         {
             var configurationAttribute = Assembly.GetExecutingAssembly().GetCustomAttribute<AssemblyConfigurationAttribute>();
+
+            if (configurationAttribute is null)
+            {
+                throw new ArgumentNullException("AssemblyConfigurationAttribute not found");
+            }
+
             if (configurationAttribute.Configuration.Equals("Debug", StringComparison.InvariantCultureIgnoreCase))
             {
                 return BuildConfiguration.Debug;
@@ -260,6 +266,11 @@ namespace Coverlet.Integration.Tests
 
         private protected void PinSDK(ClonedTemplateProject project, string sdkVersion)
         {
+            if (project is null)
+            {
+                throw new ArgumentNullException(nameof(project));
+            }
+
             if (string.IsNullOrEmpty(sdkVersion))
             {
                 throw new ArgumentException("Invalid sdkVersion", nameof(sdkVersion));
@@ -270,15 +281,19 @@ namespace Coverlet.Integration.Tests
                 throw new FileNotFoundException("coverlet.integration.template.csproj not found", "coverlet.integration.template.csproj");
             }
 
+            if (project.ProjectRootPath is null || !Directory.Exists(project.ProjectRootPath))
+            {
+                throw new ArgumentException("Invalid ProjectRootPath");
+            }
+
             File.WriteAllText(Path.Combine(project.ProjectRootPath, "global.json"), $"{{ \"sdk\": {{ \"version\": \"{sdkVersion}\" }} }}");
         }
     }
 
     class ClonedTemplateProject : IDisposable
     {
-        private bool _cleanupOnDispose;
-
-        public string? ProjectRootPath { get; private set; }
+        public string ProjectRootPath { get; private set; }
+        public bool CleanupOnDispose { get; private set; }
 
         // We need to have a different asm name to avoid issue with collectors, we filter [coverlet.*]* by default
         // https://github.com/tonerdo/coverlet/pull/410#discussion_r284526728
@@ -286,7 +301,11 @@ namespace Coverlet.Integration.Tests
         public static string ProjectFileName { get; } = "coverlet.integration.template.csproj";
         public string ProjectFileNamePath => Path.Combine(ProjectRootPath, "coverlet.integration.template.csproj");
 
-        public ClonedTemplateProject(string projectRootPath, bool cleanupOnDispose) => (ProjectRootPath, _cleanupOnDispose) = (projectRootPath, cleanupOnDispose);
+        public ClonedTemplateProject(string? projectRootPath, bool cleanupOnDispose)
+        {
+            ProjectRootPath = (projectRootPath ?? throw new ArgumentNullException(nameof(projectRootPath)));
+            CleanupOnDispose = cleanupOnDispose;
+        }
 
         public bool IsMultipleTargetFramework()
         {
@@ -314,7 +333,7 @@ namespace Coverlet.Integration.Tests
 
         public void Dispose()
         {
-            if (_cleanupOnDispose)
+            if (CleanupOnDispose)
             {
                 Directory.Delete(ProjectRootPath, true);
             }
